+++ frontend/src/components/FlowLayer.tsx
@@ -1,0 +1,1 @@
+// TODO move here the default layer from App.tsx
